# randstad/output/03-target/api-specifications.yaml
openapi: 3.0.3
info:
  title: ShopiClaw API
  version: 1.0.0
  description: >
    REST API for ShopiClaw, the agentic commerce app modernizing Shopizer v1.1.5.
    Combines commerce resources (catalog, cart, order, payment) with the embedded
    ShopiClaw agent (chat, tools, RAG). Replaces legacy Struts2 actions and JAX-WS SOAP.
    LLM provider: OpenAI. Payment via tokenization (no PAN).
servers:
  - url: /api
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: shopiclaw_session
    mcpClient:
      type: http
      scheme: bearer
      description: Per-client bearer token (OAuth client-credentials) for external MCP agents; scoped to an allow-list.
  schemas:
    Product:
      type: object
      properties:
        id: { type: string }
        sku: { type: string }
        name: { type: string }
        price: { type: integer, description: minor currency units }
        currency: { type: string, example: EUR }
        enabled: { type: boolean }
        attributes: { type: object, additionalProperties: true }
...
    Error:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
security:
  - sessionCookie: []
paths:
  /catalog/search:
    get:
      tags: [catalog]
      summary: Search catalog (legacy SearchAction; FS-0001, RULE-0007)
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: categoryId, in: query, schema: { type: string } }
      responses:
        '200':
          description: results
          content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Product' } } } }
  /catalog/products/{id}:
    get:
      tags: [catalog]
      summary: Product details (FS-0001)
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses:
        '200': { description: product, content: { application/json: { schema: { $ref: '#/components/schemas/Product' } } } }
        '404': { description: not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...